First commit from hard drive
Implementation of the ElGamal Cryptosystem. Create a file called
'Plaintext' with the message to be transferred. For this example to work
you must have the files in a read/write permission directory. It will
save the encryption key in a file, K1, and the decrption key in a file
K2. More details to come.